server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
python-multipart==0.0.9
pydantic==2.9.2
openai==1.51.0
httpx==0.27.2
